feat: add display and rules json definitions for custom credential
For the sake of learning, I created a custom credential (type `CompuComEmployee`) so that we can see what the experience would be of working with a custom credential. Docs for this are [here](https://docs.microsoft.com/en-us/azure/active-directory/verifiable-credentials/how-to-use-quickstart#create-a-custom-credential) and they include instructions for how to integrate a custom credential with an application.